requirements.txt
# ── Transcription ─────────────────────────────────────────────────────────────
# faster-whisper: CTranslate2 backend, int8 quantized, CPU-optimized
# Replaces openai-whisper (much faster on CPU, same accuracy)
faster-whisper>=1.0.0

# ── Semantic Search ────────────────────────────────────────────────────────────
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0

# ── UI ─────────────────────────────────────────────────────────────────────────
gradio>=4.44.0

# ── Audio/Video extraction ─────────────────────────────────────────────────────
ffmpeg-python>=0.2.0

# ── Utilities ──────────────────────────────────────────────────────────────────
numpy>=1.26.0
python-dotenv>=1.0.0
loguru>=0.7.0

# ── Dev / Testing ──────────────────────────────────────────────────────────────
pytest>=7.0.0
ruff>=0.1.0
